Prefer www.luxeria.ch subdomain by default
Accessing the website via the subdomain will make use the CNAME record. This means that the request is served by the GitHub CDN, leading to faster page loads. Without the subdomain, the IP address stored in the A record will be used. This can be really slow because of DoS mitigations on the GitHub server.

See also: http://helloanselm.com/2014/github-pages-redirect-performance/